Minor refactorings
Add rosbridge static singleton to avoid multiple connections to ROS
